Revert "Add new rrpack script for RR2 format"
This reverts commit 4d4ee04130c97514050c0655ac004e1176d613e6.
